input: misc: aw8624_haptic: Rename to qti-haptics
This driver behaves like the qti haptics driver and hence we can
use the qti vibrator HAL. Since [1] the qti vibrator HAL checks
wether a qti haptics device is detected and does not work if it's
not a qti one. In order to allow the HAL to work with this driver
we spoof the input device name to qti-haptics if the actual
qti-haptics driver is not built.

[1] https://github.com/LineageOS/android_vendor_qcom_opensource_vibrator/commit/c8f31f143e5ed2bfed4f47193fc7514580e87220

Change-Id: I8a233b3ea27669f80ec0a2929f5790ac36154b16
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>